+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Unreleased] 
 ### Changed
 - Read a register instead of doing a ping in the Ethernet's NetStatusListener
+- Use inspect instead of pkg_resources to find the path to the FoE application.
+
 
 ## [7.0.1] - 2023-04-04
 ### Fixed